Networking VM's together
I wanted to make it so that each VM was able to reach other other over a private network.  The biggest requirement is that the ansible machine needs to be able to reach the vanilla machine.

I also realized I had checked in .vagrant/ files but didn't really need to.  I'm removing them.