Adding ports to forward
I was following the step to install nginx and in order to get access to the webserver on the ansible server, I had to forward port 80 from the VM to 8080 on the host.  The steps also had me foward port 443 (https) to 8443.  I haven't been able to connect to the server via https://localhost:8443 but I'm not too worried.